app/requirements.txt
psycopg2==2.8.6
django==3.0.3
djangorestframework==3.12.2
requests
celery[redis]
redis
